requirements.txt
aiogram==3.25.0
aiosqlite==0.22.1
openai==2.24.0
python-dotenv==1.2.2
redis==6.4.0
fastapi==0.116.1
uvicorn==0.35.0
